requirement.txt
grpcio
grpcio-tools

Compile the Proto File:
python -m grpc_tools.protoc -I. --python_out=. --grpc_python_out=. chat.proto

Run the Server:
python server.py -client_ids 1,2,3 -port 50051

Run Clients:
python client.py -client_id 1 -server_ip localhost -port 50051 -message "Hello World!"

Run Tests:
python run_test_1.py

Build and Run with Docker:
docker build -t grpc-chat .
docker run --